+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -17,8 +17,8 @@ from demisto_sdk.commands.common.hook_validations.integration import \
 from demisto_sdk.commands.common.hook_validations.layout import LayoutValidator
 from demisto_sdk.commands.common.hook_validations.playbook import \
     PlaybookValidator
-from demisto_sdk.commands.common.hook_validations.release_notes import \
-    ReleaseNotesValidator
+from demisto_sdk.commands.common.hook_validations.old_release_notes import \
+    OldReleaseNotesValidator
 from demisto_sdk.commands.common.hook_validations.reputation import \
     ReputationValidator
 from demisto_sdk.commands.common.hook_validations.script import ScriptValidator
@@ -92,7 +92,7 @@ class TestValidators:
         (INVALID_PLAYBOOK_PATH, PLAYBOOK_TARGET, False, PlaybookValidator)
     ]
 
-    @patch.object(ReleaseNotesValidator, 'get_master_diff', return_value='Comment.')
+    @patch.object(OldReleaseNotesValidator, 'get_master_diff', return_value='Comment.')
     def test_validation_of_beta_playbooks(self, mocker):
         """
         Given
@@ -165,13 +165,13 @@ class TestValidators:
 
     INPUTS_RELEASE_NOTES_EXISTS_VALIDATION = [
         (VALID_SCRIPT_PATH, SCRIPT_TARGET, VALID_ONE_LINE_CHANGELOG_PATH, SCRIPT_RELEASE_NOTES_TARGET,
-         ReleaseNotesValidator, True),
+         OldReleaseNotesValidator, True),
         (VALID_SCRIPT_PATH, SCRIPT_TARGET, VALID_ONE_LINE_CHANGELOG_PATH, INTEGRATION_RELEASE_NOTES_TARGET,
-         ReleaseNotesValidator, False),
+         OldReleaseNotesValidator, False),
         (VALID_INTEGRATION_TEST_PATH, INTEGRATION_TARGET, VALID_ONE_LINE_CHANGELOG_PATH,
-         INTEGRATION_RELEASE_NOTES_TARGET, ReleaseNotesValidator, True),
+         INTEGRATION_RELEASE_NOTES_TARGET, OldReleaseNotesValidator, True),
         (VALID_INTEGRATION_TEST_PATH, INTEGRATION_TARGET, VALID_ONE_LINE_CHANGELOG_PATH,
-         SCRIPT_RELEASE_NOTES_TARGET, ReleaseNotesValidator, False)
+         SCRIPT_RELEASE_NOTES_TARGET, OldReleaseNotesValidator, False)
     ]
 
     @pytest.mark.parametrize('source_dummy, target_dummy, source_release_notes, target_release_notes, '
@@ -183,8 +183,8 @@ class TestValidators:
         try:
             copyfile(source_dummy, target_dummy)
             copyfile(source_release_notes, target_release_notes)
-            mocker.patch.object(ReleaseNotesValidator, 'get_master_diff', side_effect=self.mock_get_master_diff)
-            validator = ReleaseNotesValidator(target_dummy)
+            mocker.patch.object(OldReleaseNotesValidator, 'get_master_diff', side_effect=self.mock_get_master_diff)
+            validator = OldReleaseNotesValidator(target_dummy)
             assert validator.validate_file_release_notes_exists() is answer
         finally:
             os.remove(target_dummy)
@@ -216,10 +216,10 @@ class TestValidators:
             for (release_notes_file, answer) in changelog_files_answer:
                 if file_type == 'Script':
                     test_package.append((dummy_file, SCRIPT_TARGET, release_notes_file,
-                                         SCRIPT_RELEASE_NOTES_TARGET, ReleaseNotesValidator, answer))
+                                         SCRIPT_RELEASE_NOTES_TARGET, OldReleaseNotesValidator, answer))
                 elif file_type == 'Integration':
                     test_package.append((dummy_file, INTEGRATION_TARGET, release_notes_file,
-                                         INTEGRATION_RELEASE_NOTES_TARGET, ReleaseNotesValidator, answer))
+                                         INTEGRATION_RELEASE_NOTES_TARGET, OldReleaseNotesValidator, answer))
 
         return test_package
 
@@ -233,8 +233,8 @@ class TestValidators:
         try:
             copyfile(source_dummy, target_dummy)
             copyfile(source_release_notes, target_release_notes)
-            mocker.patch.object(ReleaseNotesValidator, 'get_master_diff', side_effect=self.mock_get_master_diff)
-            validator = ReleaseNotesValidator(target_dummy)
+            mocker.patch.object(OldReleaseNotesValidator, 'get_master_diff', side_effect=self.mock_get_master_diff)
+            validator = OldReleaseNotesValidator(target_dummy)
             assert validator.is_valid_release_notes_structure() is answer
         finally:
             os.remove(target_dummy)
@@ -347,7 +347,7 @@ class TestValidators:
 
     @pytest.mark.parametrize('file_path, file_type', FILE_PATHS)
     def test_is_valid_rn(self, mocker, file_path, file_type):
-        mocker.patch.object(ReleaseNotesValidator, 'get_master_diff', sreturn_value=None)
+        mocker.patch.object(OldReleaseNotesValidator, 'get_master_diff', sreturn_value=None)
         mocker.patch.object(StructureValidator, 'is_valid_file', return_value=True)
         mocker.patch.object(IntegrationValidator, 'is_valid_subtype', return_value=True)
         mocker.patch.object(IntegrationValidator, 'is_valid_feed', return_value=True)
